[Frontend] Separated signUp and codeHistory column content to LeftColumn component
